Fix: Dont check for module_type is Radio
During the extraction of Huawei GSM sites, the query
check for module_type = Radio because there are
duplicate entries with module_type = Transmission for
CORBA NBI XML dumps.

Since extraction of sites relies on a view that combines
data from the different Huawei CM format tables,
the query is empty for data other than CORBA NBI XML.

This fixes this by removing that condition and adding
distinct to remove the duplicates.